+++ vite.config.ts
@@ -1,6 +1,7 @@
 import { defineConfig } from "vite";
 import vue from "@vitejs/plugin-vue";
 import webExtension, { readJsonFile } from "vite-plugin-web-extension";
+const target = process.env.TARGET || "chrome";
 
 function generateManifest() {
   const manifest = readJsonFile("src/manifest.json");
@@ -13,22 +14,34 @@ function generateManifest() {
   };
 }
 
+function getBuildOutputDir() {
+  if (target === "firefox") {
+    return "dist/firefox";
+  } else {
+    return "dist/chrome";
+  }
+}
+
 // https://vitejs.dev/config/
 export default defineConfig({
-  build:{
-    rollupOptions:{
-      
-    }
+  define: {
+    __BROWSER__: JSON.stringify(target),
+  },
+  build: {
+    outDir: getBuildOutputDir(),
+    rollupOptions: {},
   },
   plugins: [
     vue(),
     webExtension({
       manifest: generateManifest,
-      disableAutoLaunch:true,
+      disableAutoLaunch: true,
       watchFilePaths: ["package.json", "manifest.json"],
-      additionalInputs:[
+      additionalInputs: [
         "src/offscreen/offscreen.html",
-      ]
+        "src/sidepanel/index.html",
+      ],
+      browser: process.env.TARGET || "chrome",
     }),
   ],
 });